[refactor] Fix struct identifier command sources identifiers
