+++ src/translator/commands/header/sc_scn_tex_struct_identifier_command.h
